use mmr trait to access mmr from pallet-ismp
